server/requirements.txt
Flask==2.2.2
flask_mongoengine==1.0.0
Flask_RESTful==0.3.9
pymongo==4.3.3
Werkzeug==2.2.2
uWSGI==2.0.21
celery==5.2.0
importlib-metadata==4.13.0
importlib-resources==5.12.0
kombu==5.2.4
redis==5.0.4
scipy==1.7.3

